requirements.txt
# Lean4-CodeVerifier dependencies

# LLM API client
openai>=1.0.0

# HTTP libraries
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.9.0

# Config file parsing
pyyaml>=6.0

# Azure authentication (optional; required for Azure OpenAI)
azure-identity>=1.12.0

# Logging (used by decomposer)
loguru>=0.7.0

# Progress bar
tqdm>=4.66.0
